+++ lualib-src/lua-socket.c
@@ -543,7 +543,7 @@ get_buffer(lua_State *L, int index, struct socket_sendbuffer *buf) {
 		buf->type = SOCKET_BUFFER_RAWPOINTER;
 		buf->buffer = lua_touserdata(L, index);
 		if (lua_isinteger(L, index+1)) {
-			buf->sz = lua_tointeger(L, index+1);
+			buf->sz = (size_t)lua_tointeger(L, index+1);
 		} else {
 			buf->sz = lua_rawlen(L, index);
 		}
